refactor: extract app api helpers
